Fix truncation of observing list upon adding to it from popup menu
If one added an object to the observing wishlist from the popup menu,
we used to truncate the wishlist, add the new object, and never load
the full list from disk. This would result in the entire wishlist
being lost upon exiting KStars. This commit fixes the issue by making
sure that the loading function is called even if the Observation
Planner window is not shown.

Moreover, we also improve a few UI aspects, notably we warn the user
of truncation upon canceling the loading process.

UI:
Fixes issue #163