Fix  CatalogsComponent::findByName() to find only 1 exact match
This is a fix that **might break something**, but traditionally, the assumption of `SkyMapComposite::findByName` has been to find 1 exact match. Preventing fuzzy matching will ensure that we don't report objects that are not present in our catalog as being present because an inexact match was found.

More thoroughly, since our new catalogs have object names such as "PGC 5141" in the long name, an inexact match can lead to PGC 514 being "found" in the catalog even though it doesn't exist in the database, so it is crucial to prevent hard-to-investigate bugs later.

This will ensure that, for e.g., the observing wishlist loads correctly and reports missing objects (that can later be retrieved through the internet search feature) correctly.