Bugfix initial guide deviation set blocks post MF calibration
Using the option "Only Start if Guide Deviation < xx" leads to a problem with meridian flips. The flip itself is executed, but the capture module hangs reporting that it waits for guiding deviation is below the limit. This is due to the fact that checking guide deviation is executed while guiding is still suspended after the meridian flip.
- [x] Modify the meridian flip test cases so that the problem is detected.
- [x] Shift check for initial guide limit from capture preparation to pending tasks checks