Avoid in-sequence-counter countdown when already zero
Pausing a sequence that uses in-sequence-focusing (i.e. focusing after a fixed amount of captures) creates a problem if pausing happens exactly before a meridian flip. After the meridian flip, in-sequence-focusing does not happen any more.

The reason is, that the appropriate counter is first set to 0 and then decreased, which is a bad idea for `uint`.