Bugfix: Greedy simulator needs capturedFramesCount
The greedy scheduler wasn't taking into account the previously captured frames when creating schedule estimates. That is, it would schedule and run jobs properly, but the future start/end times displayed might be poor estimates of the actual times the jobs would start/stop (with run-to-completion or repeat-N stop conditions and rememberJobProgress).

This should fix that issue. A test is added.