Prevent cpu waster from zero-delay captureDelayTimer when CCD starts slowly.
This MR avoids a flood of messages when the CCD starts slower than what Capture expects.

Timer `captureDelayTimer` is set to zero delay when starting a capture, in order to finish signal processing then proceed with waiting for the end of the exposure. If the CCD sends `CAPTURE_OK` slower than expected, the timer restarts multiple times with its zero delay and may flood the log with 200-300 messages before the CCD can post its startup message.

The change intercepts the call to `start()` with a zero-delay timer, and raises the interval a bit for the next check.