One-pulse dither should not send any guide pulses on first guide frame
Fix an issue with the one-pulse dither scheme. It seems like the first guide frame is still measuring the dithering-in-progress. Set the reticle to wherever the first guide frame finds itself after a 1-pulse dither. The way it is, it winds up often sending reasonably large guide pulses unnecessarily on the first frame.