Fix issue estimating polar-alignment target pixel when error is large
There was a hardwired search of 2 degrees in the PAA code that refreshes the current PAA error (e.g. after the user has adjusted the alt/az knobs). Thus, with large PAA error, this would fail and the error was not updated. This corrects that.

Test added.

Note that with large error there are other problems too--e.g. if the target star wanders off the image because of the adjustments. Those issues are not addressed here, and would need a change of algorithm (e.g. plate solving instead of the borrowed guiding methods that are currently used).