Bugfix for aborting guiding while dithering
When the scheduler option for target drifting is enabled, and a re-alignment is necessary, the guiding module state might get out of sync with PHD2. To test this case:
- Enable target drift and set the maximal distance to 0 so that re-alignment always happens
- Ensure that the option "Abort autoguide on failure" is disabled in the guiding options for dithering
- Enable dithering
- Start a scheduler job with tracking, alignment and guiding enabled (use PHD2 for guiding)
After the first capture and the subsequent alignment it might happen that guiding reports that guiding was already running but it doesn't. As a consequence, the scheduler is waiting infinitely long for guiding to start.