Capture state machine | introducing device adaptor
Next step extracting state handling from `Capture` into a dedicated state machine. In this step devices are managed by a dedicated device adaptor (successor of the command processor). With this refactoring it is no longer necessary to route device events from `Capture` through `SequenceJob` to the state machine, the device adaptor sends the events directly to the state machine.

Additionally, this change contains a bug fix for temperature handling.