Bugfix for PHD2 guiding - restart guiding after lost star
Restarting guiding after a lost star event can lead to the situation that `Guide` switches to the calibration state while the `PHD2` remains in the stopped stage. If in this case guiding needs to be stopped again (e.g. there is no guiding star visible), `PHD2` does not send an abort event and `Guide` remains in the calibration state.

This is fixed by this change. Additionally, a dedicated dithering state is introduced for `PHD2` to separate calibration from dithering.