Issue#103
Fixes most of the Qt deprecation warnings as flagged by Issue#103.
Of the remaining 4 warnings: 
- 1 I will raise a separate issue for - no direct migration path with Qt 6.0 but I have a proposal.
- 1 is upstream in QCustomPlot and therefore we can't fix.
- 2 I don't understand enough to fix, more to the point I don't understand how they currently work.

Also added this->done call in starhopperdialog on failure to compute a route as clicking the empty lists buttons would cause a hard crash due to list out-of-bounds.
This is my first significant MR and would benefit from careful checking. It passes unit and functional tests for me but touches a lot of places.