Geographic map update
Fix for Issue #12 
This makes the geographic map, and the dialog holding it, scaleable. It also replaces the map image with a new high resolution version from Nasa - visible earth (5400x2700px). Together these allow the user to view a much more detailed geographic map if they desire and if their display supports it.