Add support for profile scripts
This is a large MR that primarily adds support for customs delays and scripts for each driver in the equipment profile. This is sometimes required if you want to achieve some startup sequence. In my observatory, I need to turn on power first to the equipment. While this is achieved when using the scheduler, it would be nice to have it done as well when starting any equipment profile.

Sometimes a driver needs to be started before the rest of the drivers following by some delay. This MR adds support for adding pre-delay --> pre-script --> driver --> post-delay --> post script

This can be done for any driver in the equipment profile. Furthermore, now server and client managers use signals to inform Ekos on its status instead of waiting synchronously to complete the operation which can block the GUI.