Bugfix capturing does not continue after a meridian flip
This change fixes two bugs:
1. If the meridian flip is triggered at the end of a capture sequence and the scheduler has not completed the job, capturing is waiting endlessly for guiding to start. The problem is caused by a wrong condition during capture preparation which ends up that the capture preparation loop never completes.
2. If the meridian flip is triggered at the end of a sequence job and switches to the next one, it could also happen that capturing remains in the IDLE state instead of calibrating and continuing to capture. This only happens if an *initial guiding limit* is set. The problem with this limit is that during the preparation step, the capture state machine waits for a guiding deviation below this limit. If guiding has been aborted by the slewing mount of the meridian flip, the preparation state never ends.

**Test cases**

For 1., create a capture sequence with one single frame and create a scheduler job using this capture sequence, select **tracking** and **guiding** and set the job completion condition to "Repeat until terminated". Use a target close east of the meridian and start the scheduler. Ensure that the meridian flip box is checked.

The test case `TestEkosMeridianFlipSpecials::testAbortSchedulerRefocusMF()` is added to verify the fix.

After a completed meridian flip, capturing does not continue without this fix.

For 2., create a RGBRGB capture sequence where for each filter only one frame is captured. Slew to a target close east of the meridian, start guiding and start capturing. Ensure that the meridian flip box is checked.

Without this fix, `TestEkosMeridianFlip::testCaptureGuidingMF()` fails.