Allow user to increase number of SEP MultiStar reference stars, invent SNR for invented guidestars
Allows the user to increase the number of reference stars used by SEP MultiStar (using a new QSpinBox in the Guide options menu). It used to be fixed at 10 (or fewer, if too few stars were detected). It still defaults to 10, but the user can increase it up to 100. This doesn't seem to be a computational issue. When using a lot of reference stars (e.g. I experimented with 50), I found it's best not to do a lot of filtering of stars in the StellarSolver Profile Editor's star filtering parameters section for guide.

Also, fixes a display issue. The guide star position can be deduced, even if that star is not detected, by the positions of the reference stars. However, when that happened, the SNR of the guide star was set to 0 (and displayed as 0 on the SNR graph). Now, if the guide star is invented, its SNR is also estimated using the SNRs of the reference stars.