Simplify target set
Simplify target set so that whenever capture and solve is *explicitly* clicked by the user, then that indicates the user wants to solve using the current mount coords, whereever that may be. When using scheduler, this is not applicable and the last target set by the mount newTarget is kept.