Always check initial guiding deviation before capturing
If "check initial guiding deviation" is set, in some cases this check is not executed. This is partly due to a bug where the preparation step of a capture sequence does not end in the state CAPTURE_PROGRESS, which is a bug.

On the other hand, this check hasn't been executed in the middle of a capture sequence under the assumption that the normal guiding deviation check will cover these. As described in [Issue #209](https://invent.kde.org/education/kstars/-/issues/209) it makes sense to execute the initial guiding deviation before each capture.