Bugfix for initialisation of rotator device (and others)
If the rotator device shows up before the first camera device, the camera does not get the appropriate rotator set and hence the rotator button in the capture module remains disabled.

A similar case is with GPS showing up before the mount device and the weather device before the focuser.

For all these cases, the initialization is fixed.