Option added to use dedicated guiding chips with optical trains
This change was necessary due to the introduction of optical trains for cameras that bring their own guiding chip.

In order to use cameras with dedicated guiding chips, select the camera's driver in the optical train and set the guiding option "Use dedicated Guiding Chip for Guiding" (this is set by default). For cameras without a dedicated guiding chip, this option has no effect and is deactivated.
This change was necessary due to the introduction of optical trains.