PAA: use the index and healpix of the 3rd PAA calibration image at the start of the refresh phase.
This will speed up the first solve in the PAA refresh phase for the PAA plate-solve algorithm.
It assumes the solution will be similar to the 3rd calibration image's solution, and thus tries its index and healpix.
The same technique is already used for subsequent refresh image solves.
Works fine. On my mac laptop, with the simulator, the 1st refresh solve went from 1.6s to 0.2s.