Bugfix for restarting a paused capture sequence after a meridian flip
When a capture sequence is paused, the connection to the camera device
remains open. As a result, as soon as the sequence is restarted, the captured
image from the alignment is handled by mistake from the capture module instead
of leaving it to the alignment module. In cosequence, the re-alignment process
hangs and capturing will not be restarted.

This is a bug fix for issue #216.