Do not draw star circles when printing HFRs
When running "Mark Stars", if the HFR is being printed (as in the focus tab), then it's not necessary to also draw a small oval to indicate which star it is. Often those circles obscure the actual star. With this, the star can be seen alongside the HFR value.

Before this change
![zz-mark-stars-before](/uploads/bfc2cfdd57696135871b0a2ba45b5b74/zz-mark-stars-before.png)

With this change
![zz-mark-stars-after](/uploads/868af6d72eca8383d1f8205d32e6706c/zz-mark-stars-after.png)

With this change, and zoomed in
![zz-mark-stars-after-zoom](/uploads/44ad8bc7686099eb0fc9e9d9fa218bb5/zz-mark-stars-after-zoom.png)