State machine controls meridian flip for Mount
This is the next step towards dedicated state machines for EKOS. In this step the entire meridian flip control of the `Mount` is moved to the state machine.

With this step it is possible to test the state machine's meridian flip control of `Mount` separately without initializing an entire INDI infrastructure.