Control to allow autofocus backlash compensation to be toggled on or off
This small MR adds a control to allow the backlash compensation performed by the Linear and Linear 1 Pass autofocus algorithms to be toggled on or off. Currently it is always on.

Currently the Linear and Linear 1 Pass autofocus algorithms perform backlash compensation on outward focus movements. This works fine if nothing else if performing backlash compensation but may cause conflicts if the device driver or focuser itself is performing backlash compensation.

So, for example, with the ASI EAF there is no device driver compensation, not is the device itself capable of performing backlash compensation so it is appropriate to allow the Linear and L1P algorithms to perform the backlash compensation.

For the Sense Sesto 2 focuser, the device driver has been written to perform backlash compensation. It this case it makes sense to allow the driver to perform the backlash compensation and for the autofocus algorithm to not perform any compensation.

Other focusers such as the MyFocuserPro 2 have the ability to store user defined backlash parameters on the firmware of the device. In this case the user may wish to use the device to perform the backlash compensation or to switch that off and allow he focus algorithm to do this.

For more sophisticated focusers with no backlash, this MR will allow the autofocus backlash compensation to be turned off speeding up the focus process.

The "AF Backlash Comp" checkbox is defaulted on, so existing users will not need to make changes to continue getting focus results.