Bug fix for target recognition after finished slew
The update of the final mount position after a slew might come later than the mount state `MOUNT_TRACKING`. If this happens, the wrong coordinates are taken currently as new target. The side effect of this is that in some cases the meridian flip fails, because the mount slews to the wrong position.

With this change, the target position is taken from the mount coordinate updates **after** the tracking state has been received, ensuring that the mount position is correct.