Focus Motion Timer bug fix
Currently within the focus module when an Autofocus run is initiated, a timer is set in case the focuser becomes unresponsive. If the timer fires then the request to move the focuser is resent in the hope that the focuser issue is transient. If it is then autofocus should be able to successfully continue. If 3 timeouts are received then the Autofocus process attempts to abort the run.

A bug existed whereby the timer is cancelled after a position request is sent to the focuser and the focuser returns the first piece of information. Usually on a focus move indi will poll the focuser every x milliseconds to see what is happening. If the focuser is still moving then autofocus waits until the next poll period to see what's happened.

The fix resets the timer whilst the focuser is still moving and there is more information to come back.