Add healpix solver speedup to scheduler pointing check
Use the same index (image scale) and healpix (part of the sky) when repeatedly solving roughly the same part of the sky, if the internal solver is being used for the scheduler's pointing check. This significantly speeds up the 2nd through nth solving attempt (after the proper scale and healpix is determined by the first solve). If a solve fails, then the healpix and index are reset to "unknown" and a standard solve is used in the next attempt. This strategy is already successfully used in polar alignment, and ported here to the scheduler's pointing check. Tested on my laptop, solves went from 0.7s to 0.2s after the first solve. It should be more dramatic on slower hardware.

This MR also removes some old debugging printfs I noticed.