Scheduler job handling during job startup
Currently, the "aborted job management" is only active for jobs in the capture state. With this fix errors during the startup phase are also handled in the same way.

Testing it is not that easy. In an debugging environment, use manual jumps to create `return false` from functions like ` Align::captureAndSolve()`.