Scheduler warning for missing dome
If "Unpark Dome" is checked in the scheduler's observatory startup procedure, but no dome is present, the scheduler seems to hang during the startup. This is due to the fact that it waits for a dome device to appear. Instead of writing a debugging message, a log entry is created so that the user may recognize why the scheduler does not proceed.