filename Placeholders
This merge request is to continue the work of @kchoi as chronicled in #54 and !453.

The majority of the work has been manually rebasing and repairing @kchoi's work due to the time elapsed since the previous effort stalled.

Now working on a simple GUI implementation. My initial proposal for this is:
![placeholderdialog](/uploads/597f6d98291d0cd42e9538530c3d045c/placeholderdialog.png)
This is a non-modal, but always on-top when shown, dialog that is toggled by the button selectFileDirectoryB that currently launches the directory browser.

The dialog consists of a QTableWidget of tag buttons along with their description. Each tag button click adds that button into the Format QLineEdit. The extra pair of Buttons on the dialog would be: 1. The replacement button for the launching the directory browser, and 2. A insert "/" slash button to allow easy path building without having to use the keyboard.

The Dialog Cancel button would revert a stashed value to the Format QLineEdit. A new icon would probably be appropriate for the dialog open button.

Thoughts/comments?