Make the right ekos options tab display for capture and scheduler. Start options on top.
When the user is in the Ekos/Capture tab and clicks on the Options button on the bottom right, the KStars Options menu will now automatically open up to its Ekos/Capture tab. Similarly when the user is in the scheduler tab and clicks Options, the KStars Options Ekos/Scheduler sub-tab opens up. In addition, the KStars Options menu is now moved to the front when activated.