Auto focus overscan
Separated two concepts that were intertwined:
1. Backlash as specified in the Backlash field. This was used both by device drivers to adjust for backlash as well as by the Linear and Linear 1 Pass algorithms to do the out-in double movement on outward focuser movements.
2. Autofocus Overscan which is the new name for the out-in double movement as used by the Linear and L1P algorithms.

Now it is possible to set Backlash to be used by the device driver (or set to zero to disable). Note that not all drivers currently support this. Generally, the implementation in the driver adjusts the requested position by backlash ticks. This process requires an exact number of ticks.

Separately, it is now possible to set AF Overscan backlash for use by the Linear and Linear 1 Pass algorithms. Set to zero to disable. As long as this field is set >= backlash it will work. So it is not necessary to know the exact value of a focuser's backlash to use this field.

It is now possible to set both Backlash and AF Overscan. A use case for this would be a focuser with somewhat unpredictable backlash, for example, between 90 and 100 ticks. In this case set Backlash to 95 which allow reasonably accurate manual focuser movement (backlash adjusted). Note that this won't 100% compensate for all backlash every time, so by setting an AF Overscan > 5 ticks, when Linear or L1P autofocus algorithms are used, all the backlash can be compensated for.

The AF Overscan widget replaces the recently introduced AF Backlash Comp checkout which is no longer required.

The Mechanics tab of the focuser has been redesigned for the new widget.

The tab order of all the widgets in Focus has been reset.