Fix asteroid visual magnitude calculation
The calculation of visual magnitude for asteroids is not
correct. Using Skychart program as a reference one can compare
that e.g. asteroid Vesta resulted in KStars (without patch)
in value 9.18 mag vs. Skychart 8.26 mag. With the patch
the Skychart value matches the one of Skychart +/- amplitude max.
This behavior can also be observed for other asteroids.