Extract capture initialization (filters, focus on filter change) to state machine
This is the next step of extracting capture functionality from the main UI class `Capture` to a dedicated state machine. In general, preparing everything so that the camera device may start falls into three steps:
- capture preparation: set temperature, rotator angle, guiding drift, as well as special operations for  capturing flats and darks
- calibration phase: check guiding, focus, handle all the post meridian flip stuff
- capture initialization: set the correct filter and run focusing if required by filter change
After these three steps are completed successfully, the camera device may start capturing.

This change focusses on the last part, the initialization phase. With this change, the logic follows the same approach as the already completed part for capture preparation.