Avoid '+' in target names
A '+' character in target names collides with regular expression evaluation in the placeholder handling and is therefore replaced by a '_'.

Without this replacement, image counting fails for a signature holding a + as character and not as operator for regular expressions. Capturing targets like 'M81+M82' with the scheduler lead to the situation that captured frames aren't recognized. As a consequence, the counter in the filename remains fix 1.