Add a separator when generating filenames if needed
There are situations (usually handed entered or xml edited values) where the directory doesn't end with a QDir::separator() and the filename format doesn't start with one.  One was the issue in https://indilib.org/forum/ekos/13335-scheduler-endlessly-recycles-job.html#92516
This checks for that scenario, and only if necessary, adds a separator when generating full paths or signatures.