option limitSupernovaeByZoom
# Option for disabling the zoom-level based limitation for supernovae.

## Motivation

I have a python script that imports recent supernovae from [tns](https://www.wis-tns.org) into kstars, so that I only have current supernovae for observation planning.
Unfortunately due to magnitudes of SNs in the 15-20 range they are not dieplayed in wide angle views because of the zoom-level based limitation.

## Solution

I have added an option for the supernovae settings 'limit supernovae by zoom limit' which is on by default (bahavior as usual), when it is set to off, I see all supernovae independent of zoom level